requirements.txt
annotated-types==0.7.0
asgiref==3.11.1
Django==6.0.3
django-cors-headers==4.9.0
django-ninja==1.5.3
djangorestframework==3.16.1
psycopg2-binary==2.9.11
pydantic==2.12.5
pydantic_core==2.41.5
python-dotenv==1.2.2
sqlparse==0.5.5
typing-inspection==0.4.2
typing_extensions==4.15.0
